+++ mayan/apps/tags/apps.py
@@ -21,7 +21,7 @@ from mayan.apps.events.permissions import permission_events_view
 from mayan.apps.navigation.classes import SourceColumn
 
 from .events import (
-    event_tag_attach, event_tag_edited, event_tag_removed
+    event_tag_attached, event_tag_edited, event_tag_removed
 )
 from .handlers import handler_index_document, handler_tag_pre_delete
 from .html_widgets import DocumentTagWidget
@@ -85,7 +85,7 @@ class TagsApp(MayanAppConfig):
 
         ModelEventType.register(
             model=Tag, event_types=(
-                event_tag_attach, event_tag_edited, event_tag_removed
+                event_tag_attached, event_tag_edited, event_tag_removed
             )
         )