+++ mayan/apps/tags/tests/test_views.py
@@ -4,6 +4,10 @@ from mayan.apps.documents.permissions import permission_document_view
 from mayan.apps.documents.tests.base import GenericDocumentViewTestCase
 from mayan.apps.testing.tests.base import GenericViewTestCase
 
+from ..events import (
+    event_tag_attached, event_tag_created, event_tag_edited,
+    event_tag_removed
+)
 from ..links import link_tag_edit
 from ..models import Tag
 from ..permissions import (
@@ -17,6 +21,7 @@ from .mixins import DocumentTagViewTestMixin, TagTestMixin, TagViewTestMixin
 class DocumentTagViewTestCase(
     DocumentTagViewTestMixin, TagTestMixin, GenericDocumentViewTestCase
 ):
+    _test_event_object_name = 'test_tag'
     auto_upload_test_document = False
 
     def setUp(self):
@@ -26,10 +31,15 @@ class DocumentTagViewTestCase(
     def test_document_tags_list_no_permission(self):
         self._create_test_tag(add_test_document=True)
 
+        self._clear_events()
+
         response = self._request_test_document_tag_list_view()
         self.assertNotContains(
             response=response, text=force_text(s=self.test_tag), status_code=404
         )
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_tags_list_with_document_access(self):
         self._create_test_tag(add_test_document=True)
@@ -38,20 +48,30 @@ class DocumentTagViewTestCase(
             obj=self.test_document, permission=permission_tag_view
         )
 
+        self._clear_events()
+
         response = self._request_test_document_tag_list_view()
         self.assertNotContains(
             response=response, text=force_text(s=self.test_tag), status_code=200
         )
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_tags_list_with_tag_access(self):
         self._create_test_tag(add_test_document=True)
 
         self.grant_access(obj=self.test_tag, permission=permission_tag_view)
 
+        self._clear_events()
+
         response = self._request_test_document_tag_list_view()
         self.assertNotContains(
             response=response, text=force_text(s=self.test_tag), status_code=404
         )
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_tags_list_with_full_access(self):
         self._create_test_tag(add_test_document=True)
@@ -63,11 +83,16 @@ class DocumentTagViewTestCase(
             obj=self.test_tag, permission=permission_tag_view
         )
 
+        self._clear_events()
+
         response = self._request_test_document_tag_list_view()
         self.assertContains(
             response=response, text=force_text(s=self.test_tag),
             status_code=200
         )
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_trashed_document_tags_list_with_full_access(self):
         self._create_test_tag(add_test_document=True)
@@ -81,8 +106,13 @@ class DocumentTagViewTestCase(
 
         self.test_document.delete()
 
+        self._clear_events()
+
         response = self._request_test_document_tag_list_view()
         self.assertEqual(response.status_code, 404)
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_tags_list_tag_edit_link_with_full_access(self):
         # Ensure that DocumentTag instances and links are
@@ -99,6 +129,8 @@ class DocumentTagViewTestCase(
             obj=self.test_tag, permission=permission_tag_edit
         )
 
+        self._clear_events()
+
         response = self._request_test_document_tag_list_view()
 
         link_context = response.context[-1]
@@ -111,23 +143,39 @@ class DocumentTagViewTestCase(
             status_code=200
         )
 
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
+
     def test_document_tag_attach_view_no_permission(self):
         self._create_test_tag()
+
+        self._clear_events()
 
         response = self._request_test_document_tag_attach_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_tag_attach_view_with_tag_access(self):
         self._create_test_tag()
 
         self.grant_access(obj=self.test_tag, permission=permission_tag_attach)
 
+        self._clear_events()
+
         response = self._request_test_document_tag_attach_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_tag_attach_view_with_document_access(self):
         self._create_test_tag()
@@ -136,10 +184,15 @@ class DocumentTagViewTestCase(
             obj=self.test_document, permission=permission_tag_attach
         )
 
+        self._clear_events()
+
         response = self._request_test_document_tag_attach_view()
         self.assertEqual(response.status_code, 200)
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_tag_attach_view_with_full_access(self):
         self._create_test_tag()
@@ -149,10 +202,18 @@ class DocumentTagViewTestCase(
         )
         self.grant_access(obj=self.test_tag, permission=permission_tag_attach)
 
+        self._clear_events()
+
         response = self._request_test_document_tag_attach_view()
         self.assertEqual(response.status_code, 302)
 
         self.assertTrue(self.test_tag in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event.actor, self._test_case_user)
+        self.assertEqual(event.action_object, self.test_tag)
+        self.assertEqual(event.target, self.test_document)
+        self.assertEqual(event.verb, event_tag_attached.id)
 
     def test_trashed_document_tag_attach_view_with_full_access(self):
         self._create_test_tag()
@@ -164,28 +225,43 @@ class DocumentTagViewTestCase(
 
         self.test_document.delete()
 
+        self._clear_events()
+
         response = self._request_test_document_tag_attach_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
 
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
+
     def test_document_multiple_tag_attach_view_no_permission(self):
         self._create_test_tag()
+
+        self._clear_events()
 
         response = self._request_test_document_multiple_tag_multiple_attach_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_multiple_tag_attach_view_with_tag_access(self):
         self._create_test_tag()
 
         self.grant_access(obj=self.test_tag, permission=permission_tag_attach)
 
+        self._clear_events()
+
         response = self._request_test_document_multiple_tag_multiple_attach_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_multiple_tag_attach_view_with_document_access(self):
         self._create_test_tag()
@@ -194,10 +270,15 @@ class DocumentTagViewTestCase(
             obj=self.test_document, permission=permission_tag_attach
         )
 
+        self._clear_events()
+
         response = self._request_test_document_multiple_tag_multiple_attach_view()
         self.assertEqual(response.status_code, 200)
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_multiple_tag_attach_view_with_full_access(self):
         self._create_test_tag()
@@ -207,10 +288,18 @@ class DocumentTagViewTestCase(
         )
         self.grant_access(obj=self.test_tag, permission=permission_tag_attach)
 
+        self._clear_events()
+
         response = self._request_test_document_multiple_tag_multiple_attach_view()
         self.assertEqual(response.status_code, 302)
 
         self.assertTrue(self.test_tag in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event.actor, self._test_case_user)
+        self.assertEqual(event.action_object, self.test_tag)
+        self.assertEqual(event.target, self.test_document)
+        self.assertEqual(event.verb, event_tag_attached.id)
 
     def test_trashed_document_multiple_tag_attach_view_with_full_access(self):
         self._create_test_tag()
@@ -222,19 +311,29 @@ class DocumentTagViewTestCase(
 
         self.test_document.delete()
 
+        self._clear_events()
+
         response = self._request_test_document_multiple_tag_multiple_attach_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
 
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
+
     def test_document_tag_multiple_remove_view_no_permission(self):
         self._create_test_tag()
         self.test_document.tags.add(self.test_tag)
+
+        self._clear_events()
 
         response = self._request_test_document_tag_multiple_remove_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertTrue(self.test_tag in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_tag_multiple_remove_view_with_tag_access(self):
         self._create_test_tag()
@@ -242,10 +341,15 @@ class DocumentTagViewTestCase(
 
         self.grant_access(obj=self.test_tag, permission=permission_tag_remove)
 
+        self._clear_events()
+
         response = self._request_test_document_tag_multiple_remove_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertTrue(self.test_tag in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_tag_multiple_remove_view_with_document_access(self):
         self._create_test_tag()
@@ -255,10 +359,15 @@ class DocumentTagViewTestCase(
             obj=self.test_document, permission=permission_tag_remove
         )
 
+        self._clear_events()
+
         response = self._request_test_document_tag_multiple_remove_view()
         self.assertEqual(response.status_code, 200)
 
         self.assertTrue(self.test_tag in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_tag_multiple_remove_view_with_full_access(self):
         self._create_test_tag()
@@ -269,10 +378,19 @@ class DocumentTagViewTestCase(
         )
         self.grant_access(obj=self.test_tag, permission=permission_tag_remove)
 
+        self._clear_events()
+
         response = self._request_test_document_tag_multiple_remove_view()
         self.assertEqual(response.status_code, 302)
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        event = self._get_test_object_event()
+        self.assertEqual(event.actor, self._test_case_user)
+        self.assertEqual(event.action_object, self.test_tag)
+        self.assertEqual(event.target, self.test_document)
+        self.assertEqual(event.verb, event_tag_removed.id)
 
     def test_trashed_document_tag_multiple_remove_view_with_full_access(self):
         self._create_test_tag()
@@ -285,19 +403,29 @@ class DocumentTagViewTestCase(
 
         self.test_document.delete()
 
+        self._clear_events()
+
         response = self._request_test_document_tag_multiple_remove_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertTrue(self.test_tag in self.test_document.tags.all())
 
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
+
     def test_document_multiple_tag_multiple_remove_view_no_permission(self):
         self._create_test_tag()
         self.test_document.tags.add(self.test_tag)
+
+        self._clear_events()
 
         response = self._request_test_document_multiple_tag_remove_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertTrue(self.test_tag in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_multiple_tag_remove_view_with_tag_access(self):
         self._create_test_tag()
@@ -305,10 +433,15 @@ class DocumentTagViewTestCase(
 
         self.grant_access(obj=self.test_tag, permission=permission_tag_remove)
 
+        self._clear_events()
+
         response = self._request_test_document_multiple_tag_remove_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertTrue(self.test_tag in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_multiple_tag_remove_view_with_document_access(self):
         self._create_test_tag()
@@ -318,10 +451,15 @@ class DocumentTagViewTestCase(
             obj=self.test_document, permission=permission_tag_remove
         )
 
+        self._clear_events()
+
         response = self._request_test_document_multiple_tag_remove_view()
         self.assertEqual(response.status_code, 200)
 
         self.assertTrue(self.test_tag in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_document_multiple_tag_remove_view_with_full_access(self):
         self._create_test_tag()
@@ -332,10 +470,18 @@ class DocumentTagViewTestCase(
         )
         self.grant_access(obj=self.test_tag, permission=permission_tag_remove)
 
+        self._clear_events()
+
         response = self._request_test_document_multiple_tag_remove_view()
         self.assertEqual(response.status_code, 302)
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
+
+        event = self._get_test_object_event()
+        self.assertEqual(event.actor, self._test_case_user)
+        self.assertEqual(event.action_object, self.test_tag)
+        self.assertEqual(event.target, self.test_document)
+        self.assertEqual(event.verb, event_tag_removed.id)
 
     def test_trashed_document_multiple_tag_remove_view_with_full_access(self):
         self._create_test_tag()
@@ -348,21 +494,33 @@ class DocumentTagViewTestCase(
 
         self.test_document.delete()
 
+        self._clear_events()
+
         response = self._request_test_document_multiple_tag_remove_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertTrue(self.test_tag in self.test_document.tags.all())
 
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
+
     def test_tag_document_list_with_no_permission(self):
         self._create_test_tag(add_test_document=True)
 
+        self._clear_events()
+
         response = self._request_test_tag_document_list_view()
         self.assertEqual(response.status_code, 404)
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_tag_document_list_with_tag_access(self):
         self._create_test_tag(add_test_document=True)
 
         self.grant_access(obj=self.test_tag, permission=permission_tag_view)
+
+        self._clear_events()
 
         response = self._request_test_tag_document_list_view()
         self.assertContains(
@@ -374,12 +532,17 @@ class DocumentTagViewTestCase(
             status_code=200
         )
 
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
+
     def test_tag_document_list_with_document_access(self):
         self._create_test_tag(add_test_document=True)
 
         self.grant_access(
             obj=self.test_document, permission=permission_document_view
         )
+
+        self._clear_events()
 
         response = self._request_test_tag_document_list_view()
         self.assertNotContains(
@@ -390,6 +553,9 @@ class DocumentTagViewTestCase(
             response=response, text=force_text(s=self.test_document),
             status_code=404
         )
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_tag_document_list_with_full_access(self):
         self._create_test_tag(add_test_document=True)
@@ -399,6 +565,8 @@ class DocumentTagViewTestCase(
         )
         self.grant_access(obj=self.test_tag, permission=permission_tag_view)
 
+        self._clear_events()
+
         response = self._request_test_tag_document_list_view()
         self.assertContains(
             response=response, text=force_text(s=self.test_tag),
@@ -408,6 +576,9 @@ class DocumentTagViewTestCase(
             response=response, text=force_text(s=self.test_document),
             status_code=200
         )
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_tag_trashed_document_list_with_full_access(self):
         self._create_test_tag(add_test_document=True)
@@ -418,6 +589,8 @@ class DocumentTagViewTestCase(
         self.grant_access(obj=self.test_tag, permission=permission_tag_view)
 
         self.test_document.delete()
+
+        self._clear_events()
 
         response = self._request_test_tag_document_list_view()
         self.assertContains(
@@ -431,33 +604,53 @@ class DocumentTagViewTestCase(
 
 
 class TagViewTestCase(TagTestMixin, TagViewTestMixin, GenericViewTestCase):
+    _test_event_object_name = 'test_tag'
+
     def test_tag_create_view_no_permission(self):
         tag_count = Tag.objects.count()
+
+        self._clear_events()
 
         response = self._request_test_tag_create_view()
         self.assertEqual(response.status_code, 403)
 
         self.assertEqual(Tag.objects.count(), tag_count)
 
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
+
     def test_tag_create_view_with_permissions(self):
         self.grant_permission(permission=permission_tag_create)
 
         tag_count = Tag.objects.count()
+
+        self._clear_events()
 
         response = self._request_test_tag_create_view()
         self.assertEqual(response.status_code, 302)
 
         self.assertEqual(Tag.objects.count(), tag_count + 1)
 
+        event = self._get_test_object_event()
+        self.assertEqual(event.actor, self._test_case_user)
+        self.assertEqual(event.action_object, None)
+        self.assertEqual(event.target, self.test_tag)
+        self.assertEqual(event.verb, event_tag_created.id)
+
     def test_tag_delete_view_no_permission(self):
         self._create_test_tag()
 
         tag_count = Tag.objects.count()
 
+        self._clear_events()
+
         response = self._request_test_tag_delete_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertEqual(Tag.objects.count(), tag_count)
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_tag_delete_view_with_access(self):
         self._create_test_tag()
@@ -466,20 +659,30 @@ class TagViewTestCase(TagTestMixin, TagViewTestMixin, GenericViewTestCase):
 
         tag_count = Tag.objects.count()
 
+        self._clear_events()
+
         response = self._request_test_tag_delete_view()
         self.assertEqual(response.status_code, 302)
 
         self.assertEqual(Tag.objects.count(), tag_count - 1)
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_tag_multiple_delete_view_no_permission(self):
         self._create_test_tag()
 
         tag_count = Tag.objects.count()
 
+        self._clear_events()
+
         response = self._request_test_tag_delete_multiple_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertEqual(Tag.objects.count(), tag_count)
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_tag_multiple_delete_view_with_access(self):
         self._create_test_tag()
@@ -488,21 +691,31 @@ class TagViewTestCase(TagTestMixin, TagViewTestMixin, GenericViewTestCase):
 
         tag_count = Tag.objects.count()
 
+        self._clear_events()
+
         response = self._request_test_tag_delete_multiple_view()
         self.assertEqual(response.status_code, 302)
 
         self.assertEqual(Tag.objects.count(), tag_count - 1)
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_tag_edit_view_no_permission(self):
         self._create_test_tag()
 
         tag_label = self.test_tag.label
 
+        self._clear_events()
+
         response = self._request_test_tag_edit_view()
         self.assertEqual(response.status_code, 404)
 
         self.test_tag.refresh_from_db()
         self.assertEqual(self.test_tag.label, tag_label)
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_tag_edit_view_with_access(self):
         self._create_test_tag()
@@ -511,24 +724,39 @@ class TagViewTestCase(TagTestMixin, TagViewTestMixin, GenericViewTestCase):
 
         tag_label = self.test_tag.label
 
+        self._clear_events()
+
         response = self._request_test_tag_edit_view()
         self.assertEqual(response.status_code, 302)
 
         self.test_tag.refresh_from_db()
         self.assertNotEqual(self.test_tag.label, tag_label)
 
+        event = self._get_test_object_event()
+        self.assertEqual(event.actor, self._test_case_user)
+        self.assertEqual(event.action_object, None)
+        self.assertEqual(event.target, self.test_tag)
+        self.assertEqual(event.verb, event_tag_edited.id)
+
     def test_tag_list_view_with_no_permission(self):
         self._create_test_tag()
+
+        self._clear_events()
 
         response = self._request_test_tag_list_view()
         self.assertNotContains(
             response=response, text=self.test_tag.label, status_code=200
         )
 
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
+
     def test_tag_list_view_with_access(self):
         self._create_test_tag()
 
         self.grant_access(obj=self.test_tag, permission=permission_tag_view)
+
+        self._clear_events()
 
         response = self._request_test_tag_list_view()
         self.assertContains(